examples/mobc: apply rename migration(Library -> library)
