+++ examples/mobc/src/src_user/Drivers/Etc/uart_test.c
@@ -6,7 +6,7 @@
 
 #include "uart_test.h"
 #include <src_core/TlmCmd/common_cmd_packet.h>
-#include <src_core/Library/print.h>
+#include <src_core/library/print.h>
 #include "../../Settings/sils_define.h"
 #include "string.h"   // for memcpy
 #include <stdio.h>    // SILSでのprint
