+++ examples/mobc/src/src_user/TlmCmd/Ccsds/tc_transfer_frame.c
@@ -6,7 +6,7 @@
 #include "tc_transfer_frame.h"
 
 #include <string.h>
-#include <src_core/Library/crc.h>
+#include <src_core/library/crc.h>
 
 TCTF_VER TCTF_get_ver(const TcTransferFrame* tctf)
 {
