+++ examples/mobc/src/src_user/applications/driver_instances/di_aobc.c
@@ -8,11 +8,11 @@
 #include "di_aobc.h"
 #include "../../Drivers/Aocs/aobc.h"
 #include "../../TlmCmd/user_packet_handler.h"
-#include <src_core/Library/print.h>
+#include <src_core/library/print.h>
 #include <src_core/TlmCmd/common_cmd_packet_util.h>
 #include "../../Settings/port_config.h"
 #include "../../Settings/DriverSuper/driver_buffer_define.h"
-#include <src_core/Library/result.h>
+#include <src_core/library/result.h>
 
 static RESULT DI_AOBC_init_(void);
 static RESULT DI_AOBC_update_(void);
