+++ examples/mobc/src/src_user/applications/driver_instances/di_uart_test.c
@@ -5,11 +5,11 @@
  */
 #include "di_uart_test.h"
 #include <stddef.h> // for NULL
-#include <src_core/Library/print.h>
+#include <src_core/library/print.h>
 #include <src_core/TlmCmd/common_cmd_packet_util.h>
 #include "../../Settings/port_config.h"
 #include "../../Settings/DriverSuper/driver_buffer_define.h"
-#include <src_core/Library/result.h>
+#include <src_core/library/result.h>
 
 static RESULT UART_TEST_init_by_AM_(void);
 static void UART_TEST_init_(void);
